requirements.txt
apispec==3.3.0
alembic==1.4.2
Flask==1.1.2
flask-apispec==0.8.8
marshmallow==3.5.2
pyjwt==1.7.1
pytest==5.4.2
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.6
apispec-webframeworks==0.5.2